:wrench: move token parsers to dedicated files
