+++ src/autocomplete/plural.ts
@@ -1,0 +1,8 @@
+import pluralize from 'pluralize'
+
+import { Suggestor } from './token'
+
+import choice from './choice'
+
+export default (name: string): Suggestor =>
+  choice(name, pluralize(name))
